+++ src/App.js
@@ -3,13 +3,15 @@ import './App.scss';
 import { useCookies } from 'react-cookie';
 import { BrowserRouter } from 'react-router-dom';
 import { useSelector } from 'react-redux';
+import { ToastContainer } from 'react-toastify';
 
 import { SettingsContext } from './settings';
 import Header from './components/Header/Header'
 import Routes from './components/Routes'
 import GlobalNav from './components/GlobalNav/GlobalNav';
 import Footer from './components/Footer/Footer';
-import { ToastContainer } from 'react-toastify';
+import BetaBanner from './components/BetaBanner/BetaBanner';
+import BetaModal from './components/Modals/BetaModal';
 
 function App() {
   const isEmbedded = useSelector((state) => state.editor.isEmbedded);
@@ -22,9 +24,10 @@ function App() {
       
       <SettingsContext.Provider value={{theme: cookies.theme || themeDefault, fontSize: cookies.fontSize || 'small' }}>
         <BrowserRouter>
-          { isEmbedded ? null : <><GlobalNav/><Header/></> }
+          { isEmbedded ? null : <><GlobalNav/><BetaBanner/><Header/></> }
           <Routes />
           { isEmbedded ? null : <Footer/> }
+          <BetaModal/>
         </BrowserRouter>
         <ToastContainer position='bottom-center' className='toast--bottom-center' />
       </SettingsContext.Provider>